+++ cmd/corectl/dev.go
@@ -4,8 +4,10 @@ package main
 
 import (
 	"context"
+	"fmt"
 
 	"chain/core/coreunsafe"
+	"chain/core/mockhsm"
 	"chain/database/sql"
 )
 
@@ -19,4 +21,19 @@ func reset(db *sql.DB, args []string) {
 	if err != nil {
 		fatalln("error:", err)
 	}
+}
+
+func createBlockKeyPair(db *sql.DB, args []string) {
+	if len(args) != 0 {
+		fatalln("error: create-block-keypair takes no args")
+	}
+
+	hsm := mockhsm.New(db)
+	ctx := context.Background()
+	pub, err := hsm.Create(ctx, "block_key")
+	if err != nil {
+		fatalln("error:", err)
+	}
+
+	fmt.Printf("%x\n", pub.Pub)
 }
